+++ examples/the-most-complete-example/src/views/PdfjsInteraction.vue
@@ -50,7 +50,7 @@ export default {
       });
     },
     pagesRenderedHandler(pdfApp) {
-      pdfApp.pdfViewer.currentScaleValue = "page-height";
+      setTimeout(() => (pdfApp.pdfViewer.currentScaleValue = "page-height"));
     },
   },
 };